Android: Split module implementation from oldarch/newarch
